+++ examples/flashcard-block/frontend/index.js
@@ -6,7 +6,7 @@ import FlashcardContainer from './FlashcardContainer';
 import {useSettings} from './settings';
 import SettingsForm from './SettingsForm';
 
-// Determines the maximum size of the app in fullscreen mode.
+// Determines the maximum size of the extension in fullscreen mode.
 viewport.addMaxFullscreenSize({
     height: 620,
     width: 840,
@@ -18,11 +18,11 @@ viewport.addMinSize({
 });
 
 /**
- * A simple flashcard app that displays records from a chosen view.
+ * A simple flashcard extension that displays records from a chosen view.
  * Supports choosing a question field which is displayed by default and an optional answer field that
  * is hidden until shown by the user.
  */
-function FlashcardApp() {
+function FlashcardExtension() {
     const {isValid, message, settings} = useSettings();
     const [isSettingsVisible, setIsSettingsVisible] = useState(false);
     useSettingsButton(() => {
@@ -59,4 +59,4 @@ function FlashcardApp() {
     );
 }
 
-initializeBlock(() => <FlashcardApp />);
+initializeBlock(() => <FlashcardExtension />);