+++ examples/street-view-block/frontend/InstructionsView.js
@@ -163,7 +163,7 @@ export default function InstructionsView({settings}) {
                     <Fragment>
                         <Heading paddingBottom={3}>Use your existing key</Heading>
                         <Text size="xlarge">
-                            If you&apos;ve set up a map or street view app before, you may reuse the
+                            If you&apos;ve set up a map or street view extension before, you may reuse the
                             same API key. If not, you&apos;ll need to{' '}
                             <TextButton size="xlarge" onClick={() => setInstructionView(3)}>
                                 setup a new key