+++ examples/street-view-block/frontend/SettingsView.js
@@ -88,7 +88,7 @@ export default function SettingsView({settings}) {
                 {settings.table && settings.locationField ? (
                     <FormField
                         label="Geocode cache field"
-                        description="This app stores additional information for each address. Create a new text field in your table and pick it below. Other street view apps can use the same field as long as they share the same location field."
+                        description="This extension stores additional information for each address. Create a new text field in your table and pick it below. Other street view extensions can use the same field as long as they share the same location field."
                     >
                         <FieldPickerSynced
                             table={settings.table}