+++ examples/street-view-block/frontend/useSettingsStore.js
@@ -47,7 +47,7 @@ class SettingsStore {
             // If there is no value there, check the ephemaral cache,
             // which is used as a fallback when the user doesn't have
             // permission to write the the cache field. This data is
-            // lost when the app is reloaded, but will prevent repeat
+            // lost when the extension is reloaded, but will prevent repeat
             // requests to the Geocode API.
             //
             const encoded =
@@ -250,7 +250,7 @@ class SettingsStore {
 
         if (this.isMissingCacheField || this.isInvalidCacheField) {
             action =
-                'Pick a field to store cached geocoding data. This field must not be used by other map or street view apps.';
+                'Pick a field to store cached geocoding data. This field must not be used by other map or street view extensions.';
             errorKey = 'cacheFieldId';
             reason = this.isMissingCacheField
                 ? 'The geocode cache field is missing.'