+++ examples/vega-lite-block/frontend/SpecificationEditor.js
@@ -56,7 +56,7 @@ function SpecificationEditor(props) {
             // and surfaced through onSyntaxError. Since we
             // also allow invalid specification syntax, these
             // errors must be supressed.
-            // Because we cannot have an empty app here, we must
+            // Because we cannot have an empty extension here, we must
             // do something to appease to the lint gods.
             void error;
         }